+++ app/components/map/map.tsx
@@ -1,32 +1,10 @@
 "use client"
 
-import { MapContainer, TileLayer, Marker, useMapEvent } from 'react-leaflet'
+import { MapContainer, TileLayer } from 'react-leaflet'
+import { CarMarker } from '../marker'
 import "leaflet/dist/leaflet.css"
 import "./map.css"
 import L from "leaflet"
-import carIcon from "../../../public/topview_car_wagon.png";
-import { useState } from 'react'
-import getCarPosition from '../../mockApi'
-
-function CarMarker() {
-
-  const [position, setPosition] = useState<L.LatLngExpression>([35.68131,139.76716])
-
-  const ic = L.icon({
-    iconUrl: carIcon.src,
-    iconSize: [56.7, 80],
-    iconAnchor: [28, 40]
-  })
-
-  useMapEvent('click', async () => {
-    const carPosition = await getCarPosition()
-    setPosition([carPosition.lat, carPosition.lng])
-  })
-
-  return (
-    <Marker position={position} icon={ic} />
-  )
-}
 
 function Map() {
