+++ app/components/marker/index.ts
@@ -1,0 +1,1 @@
+export { default as CarMarker } from './carMarker'
